feedback implimented
